// src/app/demo/models/modelsgeneral/clienteviewmodel.ts
export class Cliente{
    clie_Id?:number;
    clie_DNI?: string;
    clie_Nombre?: string;
    clie_Apellido?:string;
    clie_CorreoElectronico?:string;
    clie_Telefono?: string;
    clie_FechaNacimiento?:string;
    clie_Sexo?:string;
    clie_DireccionExacta?: string;
    ciud_Id?: number;
    civi_Id?: number;
    usua_Creacion?: number;
    usua_Modificacion?:number;

    clie_FechaCreacion?: string;
    clie_FechaModificacion?: string;

    codigo?:string;

    civi_Descripcion?:string;
    clie_usua_Creacion?:string;
    clie_usua_Modificacion?:string;


    pais_Id?:number;
    esta_Id?:number;

    esta_Nombre?: string;
    pais_Nombre?: string;
    ciud_Descripcion?: string;
    clie_FechaCreacionn?: string;
    clie_FechaModificacionn?: string;
    clie_NombreCompleto?: string;

    
     clie_Tipo?:string;     
     tipoCliente?:string;

     cliente?: string;
}